Crate consensus

Re-exports

pub use error::ConsensusError;

pub use ledger::LedgerSubset;

Modules

error

ledger

Structs

BlockGraphExport

BootstrapableGraph

ConsensusCommandSender

ConsensusConfig

Consensus configuration Assumes thread_count >= 1, t0_millis >= 1, t0_millis % thread_count == 0

ConsensusEventReceiver

ConsensusManager

ExportActiveBlock

ExportCompiledBlock

The block version that can be exported. Note that the detailed list of operation is not exported

ExportProofOfStake

LedgerDataExport

RollCounts

RollUpdate

RollUpdates

ThreadCycleState

Enums

ConsensusCommand

Commands that can be proccessed by consensus.

ConsensusEvent

Events that are emitted by consensus.

DiscardReason

ExportBlockStatus

Block status in the graph that can be exported.

Status

Functions

start_consensus_controller

Creates a new consensus controller.