Trait consensus::consensus_provider::ConsensusProvider

pub trait ConsensusProvider {
    fn start(&mut self) -> Result<()>;

    fn stop(&mut self);
}

Public interface to a consensus protocol.

Required methods

fn start(&mut self) -> Result<()>

Spawns new threads, starts the consensus operations (retrieve txns, consensus protocol, execute txns, commit txns, update txn status in the mempool, etc). The function returns after consensus has recovered its initial state, and has established the required connections (e.g., to mempool and executor).

fn stop(&mut self)

Stop the consensus operations. The function returns after graceful shutdown.

Loading content...